requirements.txt
flask
python-dotenv
bcrypt
#flask-restplus==0.9.2
Flask-SQLAlchemy==2.1
Werkzeug==0.16.1
aniso8601==6.0.0
attrs==19.1.0
click==7.0
flask-restplus==0.12.1
#flask==1.0.3
itsdangerous==1.1.0
jinja2==2.10.1
jsonschema==3.0.1
markupsafe==1.1.1
pyrsistent==0.15.2
pytz==2019.1
six==1.12.0
boto3
#werkzeug==0.15.4
pymysql